Added tests, benchmarks and improved performance.
More tests to actually check if we return the shortest path.
Added benchmarks to see the difference between the path and its reverse.
Exposed ReadWords(...) to the package API and added the length
constraint. The length constraint allows us to read only the words that
are of the length we care about, so even the naive way we still build
the adjacency list/graph is fast enough.
Added wordlist.txt to the testdata, it contains a sizable amount of
words.
And a smaller file to test ReadWords(...).